+++ src/Jieba/CacheFactory.php
@@ -5,6 +5,7 @@ namespace Jieba;
 use Cache\Adapter\Common\AbstractCachePool;
 use Cache\Adapter\PHPArray\ArrayCachePool;
 use Closure;
+use Jieba\Option\Dict;
 
 class CacheFactory
 {
@@ -30,6 +31,25 @@ class CacheFactory
      * @var AbstractCachePool
      */
     protected static $cachePool;
+
+    /**
+     * @param AbstractCachePool $cachePool
+     * @param Dict $dict
+     * @param string|null $fileType
+     * @return mixed
+     * @throws Exception
+     * @todo use more meaningful cache key instead of md5 values.
+     */
+    public static function getDict(AbstractCachePool $cachePool, Dict $dict, $fileType = null)
+    {
+        return self::get(
+            $cachePool,
+            md5($dict->getDictBaseName($fileType)),
+            function () use ($dict, $fileType) {
+                return $dict->getDictFileContent($fileType);
+            }
+        );
+    }
 
     /**
      * @param AbstractCachePool $cachePool